refactored the FollowersPresenter
