+++ app/src/main/java/edu/byu/cs/tweeter/client/presenter/FollowersPresenter.java
@@ -3,110 +3,22 @@ package edu.byu.cs.tweeter.client.presenter;
 import edu.byu.cs.tweeter.client.cache.Cache;
 import edu.byu.cs.tweeter.client.model.service.FollowService;
 import edu.byu.cs.tweeter.client.model.service.UserService;
-import edu.byu.cs.tweeter.client.model.service.handler.observer.GetUserObserver;
 import edu.byu.cs.tweeter.client.model.service.handler.observer.PagedNotificationObserver;
+import edu.byu.cs.tweeter.model.domain.AuthToken;
 import edu.byu.cs.tweeter.model.domain.User;
 
 import java.util.List;
 
-public class FollowersPresenter {
-    private static final int PAGE_SIZE = 10;
-
-    public interface View {
-        void displayErrorMessage(String message);
-
-        void setLoadingStatus(boolean value);
-
-        void addFollowers(List<User> followers);
-
-        void startActivity(User user);
-    }
-
-    private View view;
+public class FollowersPresenter extends PagedPresenterUser {
     private FollowService followService;
-    private UserService userService;
-
-    private User lastFollower;
-
-    private boolean hasMorePages;
-    private boolean isLoading = false;
 
     public FollowersPresenter(View view) {
-        this.view = view;
+        super(view);
         followService = new FollowService();
-        userService = new UserService();
     }
 
-    public boolean hasMorePages() {
-        return hasMorePages;
-    }
-
-    public void setHasMorePages(boolean hasMorePages) {
-        this.hasMorePages = hasMorePages;
-    }
-
-    public boolean isLoading() {
-        return isLoading;
-    }
-
-    public void setLoading(boolean loading) {
-        isLoading = loading;
-    }
-
-    public void loadMoreItems(User user) {
-        if (!isLoading) {   // This guard is important for avoiding a race condition in the scrolling code.
-            setLoading(true);
-            view.setLoadingStatus(isLoading());
-            followService.getFollowers(Cache.getInstance().getCurrUserAuthToken(), user, PAGE_SIZE, lastFollower, new GetFollowersObserver());
-        }
-    }
-
-    public class GetFollowersObserver implements PagedNotificationObserver<User> {
-
-        @Override
-        public void handleSuccess(List<User> followers, boolean hasMorePages) {
-            setLoading(false);
-            view.setLoadingStatus(isLoading());
-            lastFollower = (followers.size() > 0) ? followers.get(followers.size() - 1) : null;
-            setHasMorePages(hasMorePages);
-            view.addFollowers(followers);
-        }
-
-        @Override
-        public void handleMessage(String message) {
-            setLoading(false);
-            view.setLoadingStatus(isLoading());
-            view.displayErrorMessage("Failed to get following: " + message);
-        }
-
-        @Override
-        public void handleException(Exception exception) {
-            setLoading(false);
-            view.setLoadingStatus(isLoading());
-            view.displayErrorMessage("Failed to get following because of exception: " + exception.getMessage());
-        }
-    }
-
-    private class GetUserObserver implements edu.byu.cs.tweeter.client.model.service.handler.observer.GetUserObserver {
-
-        @Override
-        public void handleSuccess(User user) {
-            view.startActivity(user);
-        }
-
-        @Override
-        public void handleMessage(String message) {
-            view.displayErrorMessage("Failed to get user's profile: " + message);
-        }
-
-        @Override
-        public void handleException(Exception ex) {
-            view.displayErrorMessage("Failed to get user's profile because of exception: " + ex.getMessage());
-        }
-
-    }
-
-    public void getUser(String userAlias) {
-        userService.getUser(Cache.getInstance().getCurrUserAuthToken(), userAlias, new GetUserObserver());
+    @Override
+    public void getItems(AuthToken authToken, User user, int pageSize, User lastFollower) {
+        followService.getFollowers(authToken, user, pageSize, lastFollower, new GetFollowersObserver());
     }
 }
