+++ app/src/main/java/edu/byu/cs/tweeter/client/presenter/Presenter.java
@@ -1,5 +1,5 @@
 package edu.byu.cs.tweeter.client.presenter;
 
-public abstract class PagePresenter {
+public abstract class Presenter {
 
 }
